move SingleProcessMessage test to core as ProcessMessage.test.ts
